+++ mobile/android/android-components/components/browser/engine-gecko/src/main/java/mozilla/components/browser/engine/gecko/GeckoEngine.kt
@@ -93,7 +93,7 @@ typealias NativePermissionPromptResponse = org.mozilla.geckoview.WebExtension.Pe
 /**
  * Gecko-based implementation of Engine interface.
  */
-@Suppress("LargeClass", "TooManyFunctions")
+@Suppress("LargeClass", "TooManyFunctions", "DEPRECATION")
 class GeckoEngine(
     context: Context,
     private val defaultSettings: Settings? = null,